chore: refactor address-book for context awareness
